# swagger.json
{
    "swagger": "2.0",
    "basePath": "/api",
    "servers": [
        "https:\/\/nurutomo.herokuapp.com\/",
        "http:\/\/localhost:5000"
    ],
    "paths": {
        "/ssweb": {
            "get": {
                "summary": "Returns Screenshot Image from Website",
                "operationId": "get_ssweb",
                "parameters": [{
                    "description": "Absolute URL",
                    "name": "url",
                    "type": "string",
                    "in": "query",
                    "required": true
                }, {
                    "description": "Full Page Screenshot",
                    "name": "full",
                    "type": "boolean",
                    "enum": [true, false],
                    "in": "query"
                }, {
                    "description": "Delay beetween page loaded and screenshot",
                    "name": "delay",
                    "type": "integer",
                    "minimal": 0,
                    "in": "query"
                }, {
                    "description": "Result Image Format",
                    "name": "type",
                    "type": "string",
                    "enum": ["png", "jpeg"],
                    "in": "query",
                    "default": "png"
                }, {
                    "description": "JPEG Image Quality",
                    "name": "quality",
                    "type": "float32",
                    "minimal": 0.0,
                    "maximal": 1.0,
                    "in": "query"
                }],
                "responses": {
                    "200": {
                        "description": "Successfully Sceeenshoted",
                        "content": {
                            "image/*": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid URL",
                        "content": {
                            "application/json": {
                                "type": "string",
                                "format": "string"
                            }
                        }
                    }
                },
                "tags": ["Chrome"]
            }
        },
        "/canvas": {
            "post": {
                "summary": "Returns Screenshot Image from Website",
                "operationId": "get_ssweb",
                "consumes": "text/plain",
                "parameters": [{
                    "description": "Your javascript code with canvas element\nVariables:\n- c (CanvasDOMElement)\n- ctx (Canvas Context)",
                    "name": "code",
                    "type": "string",
                    "in": "body",
                    "default": "c.width = c.height = 512;\n\nctx.fillStyle = \"#0099FF\";\nctx.fillRect(128, 128, 128, 128);"
                }, {
                    "description": "Result Image Format",
                    "name": "type",
                    "type": "string",
                    "enum": ["png", "jpeg", "webp"],
                    "in": "query",
                    "default": "png"
                }, {
                    "description": "JPEG Image Quality",
                    "name": "quality",
                    "type": "float32",
                    "minimal": 0.0,
                    "maximal": 1.0,
                    "in": "query"
                }],
                "responses": {
                    "200": {
                        "description": "Successfully Drawing",
                        "content": {
                            "image/*": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "tags": ["Chrome"]
            }
        }
    },
    "info": {
        "title": "Nurutomo API",
        "version": "1.0",
        "description": "Media manipulation API"
    },
    "produces": ["image/*", "application/json"],
    "consumes": ["plain/text"],
    "tags": [{
        "name": "Chrome",
        "description": "This API uses Chromium to work (Puppeteer)"
    }],
    "responses": {
        "501": {
            "description": "Internal Server Error"
        },
        "ParseError": {
            "description": "Cannot parse this"
        },
        "MaskError": {
            "description": "There was an error"
        }
    }
}